// src/app/domain/situacao-solicitacao/situacao-solicitacao.enum.ts
export enum SituacaoSolicitacaoEnum {
  INICIADA = 'INICIADA',
  DEVOLVIDA = 'DEVOLVIDA',
  PARA_APROVAR = 'PARA_APROVAR',
  REPROVADA = 'REPROVADA',
  CANCELADA = 'CANCELADA',
  PARA_VERIFICAR = 'PARA_VERIFICAR',
  VERIFICANDO = 'VERIFICANDO',
  VERIFICADA = 'VERIFICADA',
  APROVADA_PARA_AUTORIZAR = 'APROVADA_PARA_AUTORIZAR',
  EM_ANALISE = 'EM_ANALISE',
  ANALISADA = 'ANALISADA',
  NAO_AUTORIZADA = 'NAO_AUTORIZADA',
  PARA_INSTRUCAO = 'PARA_INSTRUCAO',
  INSTRUINDO = 'INSTRUINDO',
  INSTRUIDA = 'INSTRUIDA',
  AUTORIZADA_PARA_CALCULAR = 'AUTORIZADA_PARA_CALCULAR',
  CALCULANDO = 'CALCULANDO',
  CALCULADA_PARA_PAGAR = 'CALCULADA_PARA_PAGAR',
  REALIZANDO_PAGAMENTO = 'REALIZANDO_PAGAMENTO',
  PAGAMENTO_REALIZADO = 'PAGAMENTO_REALIZADO',
  CONFERINDO_PAGAMENTO = 'CONFERINDO_PAGAMENTO',
  PARA_CORRIGIR_INFORMACOES_PAGAMENTO = 'PARA_CORRIGIR_INFORMACOES_PAGAMENTO',
  INFORMACOES_PAGAMENTO_CORRIGIDAS = 'INFORMACOES_PAGAMENTO_CORRIGIDAS',
  PAGAMENTO_CONFORMIDADE = 'PAGAMENTO_CONFORMIDADE',

  INICIADA_LABEL = 'Iniciada',
  DEVOLVIDA_LABEL = 'Devolvida',
  PARA_APROVAR_LABEL = 'Para aprovar',
  REPROVADA_LABEL = 'Reprovada',
  CANCELADA_LABEL = 'Cancelada',
  PARA_VERIFICAR_LABEL = 'Para verificar',
  VERIFICANDO_LABEL = 'Verificando',
  VERIFICADA_LABEL = 'Verificada',
  APROVADA_PARA_AUTORIZAR_LABEL = 'Aprovada/Para autorizar',
  EM_ANALISE_LABEL = 'Em análise',
  ANALISADA_LABEL = 'Analisada',
  NAO_AUTORIZADA_LABEL = 'Não autorizada',
  PARA_INSTRUCAO_LABEL = 'Para instrução',
  INSTRUINDO_LABEL = 'Instruindo',
  INSTRUIDA_LABEL = 'Instruída',
  AUTORIZADA_PARA_CALCULAR_LABEL = 'Autorizada/Para calcular',
  CALCULANDO_LABEL = 'Calculando',
  CALCULADA_PARA_PAGAR_LABEL = 'Calculada/Para pagar',
  REALIZANDO_PAGAMENTO_LABEL = 'Realizando pagamento',
  PAGAMENTO_REALIZADO_LABEL = 'Pagamento realizado',
  CONFERINDO_PAGAMENTO_LABEL = 'Conferindo pagamento',
  PARA_CORRIGIR_INFORMACOES_PAGAMENTO_LABEL = 'Para corrigir informações de pagamento',
  INFORMACOES_PAGAMENTO_CORRIGIDAS_LABEL = 'Informações de pagamento corrigidas',
  PAGAMENTO_CONFORMIDADE_LABEL = 'Pagamento em conformidade'
}